feat: Enable AI assistant only for SAP users (#3727)
* feat: Enable AI assistant only for SAP users

* adjust jsonata fn
